requirements.txt
requests
beautifulsoup4
fastapi
openai
uvicorn
redis
python-dotenv